restructure: move steps
